More compiler warning fixes. Switched automatic culling without camera off as suggested on the forum.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@918 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/irrMath.h
@@ -10,13 +10,13 @@
 #include <math.h>
 
 #if defined(_IRR_SOLARIS_PLATFORM_) || defined(__BORLANDC__) || defined (__BCPLUSPLUS__)
-	#define sqrtf(X) sqrt(X)
-	#define sinf(X) sin(X)
-	#define cosf(X) cos(X)
-	#define ceilf(X) ceil(X)
-	#define floorf(X) floor(X)
-	#define powf(X,Y) pow(X,Y)
-	#define fmodf(X,Y) fmod(X,Y)
+	#define sqrtf(X) (f32)sqrt(X)
+	#define sinf(X) (f32)sin(X)
+	#define cosf(X) (f32)cos(X)
+	#define ceilf(X) (f32)ceil(X)
+	#define floorf(X) (f32)floor(X)
+	#define powf(X,Y) (f32)pow(X,Y)
+	#define fmodf(X,Y) (f32)fmod(X,Y)
 #endif
 
 namespace irr
@@ -234,7 +234,7 @@ namespace core
 
 	inline f32 round_( f32 x )
 	{
-		return (f32)floor( x + 0.5f );
+		return floorf( x + 0.5f );
 	}
 
 	REALINLINE void clearFPUException ()
